Removed DirectX Install
FOR SOME REASON it now pops up a window asking the user if it wants to restart the computer after install. EVEN IF it's being installed from Winget. EVEN IF it's being installed "unattended". Last i checked, "unattended" means "it doesn't f**king hang waiting for the user to tell it if it wants to restart the computer or not".